Core installable: part 1, install
Begins migration of core from the old un/installer to the new
installables API. We do it in parts, to make it more clear what code is
being moved where, and what is being modified in the process.

The install code is fairly simple, and no real modifications of the
install routine itself is done in the process; the code is just
shuffled around so that it is run with the new installables API.

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404