Core installable: part 2, uninstall
Part of migration of core from the old un/installer to the new
installables API. We do it in parts, to make it more clear what code is
being moved where, and what is being modified in the process.

The uninstall code is fairly straightforward, and the main modification
that has been made in the process is that the components uninstall code
has been updated to utilize the new installables API.

Other than that, most changes are just around the migration, and don't
actually affect the underlying behavior. One minor change is that we
now autoload the core classes during uninstall, which should be safe,
since no unexpected code should be run inside of an autoloaded class
file (the main reason that we don't just load the whole plugin during
uninstall, as an extra precaution).

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404