Core installable: part 3, update
Part of migration of core from the old un/installer to the new
installables API. We do it in parts, to make it more clear what code is
being moved where, and what is being modified in the process.

The update code has been modified significantly during the process,
although for the most part the actual behavior has not been changed.

One thing worthy of note is that the update routines for versions 1.3.0,
1.5.0, and part of 2.4.0-alpha-2 and alpha-3 were all running the code
to install the custom caps. This became more obvious with the condensed
design of the new API. As a result, we realized that it isn't actually
necessary to do this, we can just run this when updating to the most
recent version that needs it (2.4.0-alpha-3). This will mean that if
we're updating from a version lower than those other versions, the
update will still get triggered as expected (since it is higher than
them).

The code relating to deactivating an extension that has been merged into
core has also been tweaked, to reduce duplication between the code for
single sites and the code for the network on multisite. For the per-site
deactivation, we can just use a callback routine with the activation
function (there's no need to check if the extension is active as was
done before, since the deactivation function does that internally).

See #404