Points installable: part 1, install
Begins migration of the points component from the old un/installer to
the new installables API. We do it in parts, to make it more clear what
code is being moved where, and what is being modified in the process.

The install code is fairly simple, but we've made two significant
changes:

- The `'wordpoints_default_points_type'` option is no longer added,
  since `update_option()` now enables autoloading fo the option by
  default as of WP 4.2. This code was added before that, see b68c02b.
- Instead of removing all capabilities on install (or re-activation,
  specifically) we now only remove the one that is affected by network-
  wide activation status.

Beside that, the changes are just porting the code over to the new API
without actually changing behavior.

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404