Points installable: part 2, uninstall
Part of migration of core from the old un/installer to the new
installables API. We do it in parts, to make it more clear what code is
being moved where, and what is being modified in the process.

The uninstall code is fairly straightforward, and mainly the code has
just been migrated over.

The one thing that is worthy of note is that we now automatically set up
autoloading for components during uninstall, if they use the installable
API. This only works for core components, of course. (This doesn't
affect extensions because it is done after they have been uninstalled.)

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404