Points installable: part 3, update
Part of migration of core from the old un/installer to the new
installables API. We do it in parts, to make it more clear what code is
being moved where, and what is being modified in the process.

The update code has undergone some modifications during the migration,
although for the most part the underlying behavior is not affected.

The main change that is worthy of note is that the update for 1.5.1 has
been entirely omitted. This update intended to set the correct charset
and collation on the DB tables, since we omitted to set them when they
were created. It did this correctly at the time, however, it is
potentially risky since we don't know whether everything will be
converted properly from whatever the current charset is to whatever the
database used by default. As such, and since there is nobody on this
ancient version of the plugin still anyway, we've decided to remove that
update. This does mean that the 2.0.0 update that moves to utf8mb4 will
probably not update those tables (in the hypothetical situation that
someone installed the old version and updated from it), because it only
updates the tables if they are already utf8.

See #404