Ranks installable: part 1, install
Begins migration of the ranks component from the old un/installer to
the new installables API. We do it in parts, to make it more clear what
code is being moved where, and what is being modified in the process.

The install code changes are just porting the code over to the new API
without actually changing behavior.

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404