Ranks installable: part 2, uninstall
Part of migration of the ranks component from the old un/installer to
the new installables API. We do it in parts, to make it more clear what
code is being moved where, and what is being modified in the process.

The uninstall code is pretty simple, and the code has just been migrated
over.

No unit tests are really of much use here; this is all tested by the
uninstall tests.

See #404