Ranks installable: part 3, update
Part of migration of the ranks component from the old un/installer to
the new installables API. We do it in parts, to make it more clear what
code is being moved where, and what is being modified in the process.

The code has just been migrated to the new API, the underlying behavior
is not affected.

The tests for the 2.0.0 update had to be modified because we can no
longer just run the update for a particular version in that manner. We
could have simply just run the update routine in question, rather than
calling upon the installable class as a whole, but we want to be as
realistic as possible.

See #404