Core installable: ensure all options removed on uninstall
There were two options that were not being deleted under the new API:

- `'wordpoints_installable_versions'`, which was just introduced as part
  of the new API.
- `'wordpoints_data'`, which has been around since day one and was
  previously handled correctly.

Removing the former is easy, we just needed to add it to the list. For
the latter, however, a few changes needed to be made.

The reason that the option was lingering even though it is already on
the list of options to uninstall, is that it was getting recreated when
core's `unset_db_version()` method was called. That's because the
design of the parent method is to pull up an array option, which causes
an empty array to be returned even when the value is deleted, and this
was then being saved. To fix that we just override the parent method
with an empty one.

There was another piece to this, however, because we found that the
option was still not being removed on multisite, for similar reasons.
Except that in this case it wasn't core's db version that was being
unset, of course, since we'd already added the blank method. It turned
out that it was actually extensions. We realized that the per-site
uninstall routines for extensions and components were actually being
run *after* core's per-site uninstall routines, because they were
necessarily offered as network uninstall routines. Thus, after core had
deleted the option on each site, the extension/component uninstall would
trigger it to be added again as an empty array in the manner described
above.

We determined that the best way to fix this was to run the extension and
component uninstaller routines manually within the uninstall routine
getter function. They are supposed to be run before any of core's
routines are run anyway. In the future, we'd have the option to just get
all of the extension and component routines and prepend them to the list
we return for core, thus bundling them all together. However, we can't
currently do that because not all extensions are guaranteed to be using
the new API yet, and so will have an un/installer instead. We can't run
per-site routines for them ourselves in that case, we just have to
trigger their un/installer to run once. So we run the extensions and
components uninstallers manually for now.

See #404