Uninstaller: don't query all sites unnecessarily
When we are network installed the list of installed site IDs will be the
list of all sites, so if the network is large we already know that it
will fail this test. So rather than running the query, we can just check
if we are network installed.

We still need to check the list of sites when we aren't network
installed though.

See #404, #714