Legacy installable shim: run updates per-site when not network active
When an extension was not network active, the old installables class
would still run the updates for all sites at once. However, in the new
installables app this was fixed for the new installable objects. We
still weren't handling it properly in our shim for legacy entities still
using the un/installer API though, so we've updated it to pass in the
`$network` parameter correctly for extensions.

See #706, #404