Updater routine factory class
Introduces a factory class for update routines, to enable most logic to
be removed from the installables an into the updater itself.

The installable must now return an array of update routine factories to
the updater, rather than the update routines themselves. The updater
then uses the factories to retrieve the routines that it determines need
to run.

The logic in the updater is also updated to determine the network
version for per-site installed entities separately from per-site
versions. The installer is updated to save the per-site version in that
case as well, and not just when network-active.

The core installables have been updated to use the new factory design.

Because one of the updater objects needs to be constructed with 5 args,
the utility function for that purpose had to be updated to support 5
args instead of just 4.

Fixes #706
See #404