Updater: correctly detect network version for legacy installables
Legacy installables that used to use the un/installer API did not set
the network version when per-site installed, so we have to detect it
based on the per-site versions.

While we are at it, we also correctly set the per-site versions to match
this version.

I just happened to think that this could have gone in an update routine
for legacy installables to use rather than baking it into the API, but
I doubt that would have worked since we need this in order to know which
updates to even run. :-)

See #706