Mock installables app during PHPUnit tests
Fixes issues resulting from the installables app not being mocked during
some tests, leading to the installable remaining registered with the
app, and perpetuating its version in the database. In order to fully fix
this, we needed to also allow the installable to be registered more than
once, so that it would have its app registered again. Some issues with
the tests were also revealed by this, and fixed.

We also realized that we use 'extension' as the type for the
installables registered in the new code, but the legacy code was still
using 'module', so we've standardized on 'extension'.

See #404