phpunit: use longer periods in hook periods extension tests
These periods are absolute, meaning that they are subject to the clock.
When the period ends with respect to the clock (the next minute comes),
the period is over. This can happen during the test, but it is less
likely to happen if we use a longer period, like one day, rather than
just one minute. (That's a 3600x reduction in chances of hitting this
when the test is running.)

We were occasionally experiencing random failures of some of these
tests, like:

```
1) WordPoints_Hook_Extension_Periods_Test::test_should_hit_period_started_negative_offset with data set "child" (array(array(array(array(60, array(array('test_entity', 'child')))))), array(array('8fd6eba3474b89832e4f275bba034...834828')))

Failed asserting that actual size 2 matches expected size 1.
```

This should now prevent that.