Hooks JS: only allow conditions on entities that are enumerable
Previously we disallowed on all top-level entities and allowed on all
sub-entities, like in an entity array. However, this doesn't actually
reflect whether or not we support conditions for them, other than
conditions where the user would have to enter the ID. The change
corrects this so that now we filter out all entities that we can't offer
conditions on, and leave all those that we do.

Fixes #718