Entities: allow ID to be specified as being an integer
When used in the GUID of the entity, the type of the ID as either an
integer or string makes a difference, particularly when serialized as
JSON and then queried against in the database.

By letting entities with integer IDs declare this fact, we can ensure
that when the object returns an ID it is in fact always an integer. (It
is easy for us to be supplied with a string pulled directly from the DB,
for example.)

We now do this for our core entities for which this is true, the post,
comment, term, and user entities.

We also include the necessary routines to update the values in the DB to
be integers, if they are not. In core, these values are only stored as
serialized JSON in the GUIDs of entities in the hook hits and points log
meta tables. Update routines are included to fix each of these, and will
be run for the core entities that need it on update. These are all of
post and comment entities, and the user entity. The term entities don't
need to be corrected, as they have just been introduced.

The update routines can also be used by extensions that need to.

Fixes #556