Hooks JS: only disallow equals conditions on non-enumerable entities
Previously we disallowed all conditions on them. The difference wasn't
noticeable because there are no other conditions for entities in core.

We also update the functions for getting data about conditions to return
a clone, so that when we modify it we are modifying the copy and not the
data itself.

See #718