Entities API: don't rely on current context in relationship class
In #535 it became possible to retrieve entities from a different context
than the entity actually exists in, by passing in the full GUID of the
entity. Switching to the proper context to retrieve the entity was now
handled internally.

However, when retrieving entity children, like a related entity, the
current context was still always assumed to be correct. This commit
fixes that in regard to relationships, by:

- Ensuring that the correct context is switched to when retrieving the
related entity IDs, and
- Passing in the full GUIDs to the related entity object when its value
is set, rather than just the IDs, so that it can also switch to the
correct context to retrieve the entity object (as #535 made possible).

The base implementation here only covers the most common form of
relationship, where the related entity comes from a context the same as
or above the context of the primary entity. For relationships where the
related entity comes from a lower or different context, special handling
would be needed to fully support out-of-context retrieval of the related
entity.

See #536