Entities API: don't rely on current context in attribute class
In #535 it became possible to retrieve entities from a different context
than the entity actually exists in, by passing in the full GUID of the
entity. Switching to the proper context to retrieve the entity was now
handled internally.

However, when retrieving entity children, like an entity attribute, the
current context was still always assumed to be correct. This commit
fixes that in regard to attributes, by ensuring that the correct context
is switched to when retrieving the attribute value.

The base implementation here only covers the most common form of
attribute, where the attribute value comes from the same context as the
entity. For attributes where the value comes from a different context,
special handling would be needed to fully support out-of-context
retrieval of the attribute value. Though this situation is unlikely to
exist in practice.

See #536