Uninstall: don't call extension deactivation function
Instead just use a filter that we introduce (
`'is_wordpoints_extension_active`).

This is better, because the deactivation only takes care of the current
site on multisite anyway.

It also avoids having to call `is_wordpoints_network_active()` within or
triggered by that function, which is unreliable during uninstall.

See #399, #712