Components: register installable on load() instead of register()
The installable only needs to be registered for active components, and
doing it here saves us a duplicate `is_active()` check in `register()`,
as well as making it possible to pull up the installables app just once,
which also improves efficiency.

Most importantly though, it avoids having to call
`is_wordpoints_network_active()` (which is needed to determine which
option to retrieve to get the list of active components, in
`register()`. This is a problem because we need to register the
components during uninstall, but checking activation status is
unreliable during uninstall.

See #399