Fix extension uninstall for installables
The code for uninstalling an extension with an installable class was
broken because we were passing the basename of the extension to
`wordpoints_get_module_data()`, when it actually expects the full path.
Our tests were updated in the last commit so that this is now tested,
and of course that exposed that it wasn't working. We thought it was
worth committing the fix separately, since it wasn't directly related
to the main brunt of that commit.

See #404