Core: prevent updates when the PHP version is incompatible
Previously we only disabled the related UI elements, we didn't actually
hook into the updater.

Now that we do, we can remove the hack that used JS to disable the
checkbox that would be used for bulk update. The checkbox should still
be able to be used for other bulk operations anyway.

This not only removes the use of inline scripts here, but for added
bonus points we also remove a use of `md5()` (which should be avoided,
just to set a good example).

See #582