Remove points info from the Profile admin screen
The points can now be edited via the new User Points screen, so the old
functionality is being moved to an extension instead. We show a notice
offering for users to install the extension if they want to continue to
use the old feature.

Removes the final uses of inline styles in core. See #582.

See #48
Fixes #579