hooks: let reactors validate a target before firing
A reactor is not always able to react to a target if it does not meet
certain requirements. For example, if the current user is actually a
guest and not a logged in user, the Points reactor cannot award points.
By letting the reactor validate the target, we are improving performance
by preventing an unnecessary fire from being processed by the
extensions, as well as ensuring that unusable data is not added to the
database. In particular, this will improve performance of the Visit
event, by avoiding calling the Periods extension into play. This is
especially important because the Visit event can be triggered on every
single page-load.

See #762, #640